refactor : 공통(global) 패키지 이동
