+++ src/main/java/likelion/festival/booth/domain/Booth.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 import com.sun.istack.NotNull;
 import likelion.festival.comment.domain.Comment;
-import likelion.festival.entity.*;
+import likelion.festival.global.image.domain.Image;
 import likelion.festival.like.domain.Likes;
 import likelion.festival.menu.domain.Menu;
 import lombok.AllArgsConstructor;
