+++ src/main/java/likelion/festival/comment/appliction/CommentService.java
@@ -6,12 +6,12 @@ import likelion.festival.comment.appliction.dto.CommentRequestDto;
 import likelion.festival.comment.appliction.dto.CommentResponseDto;
 import likelion.festival.booth.domain.Booth;
 import likelion.festival.comment.domain.Comment;
-import likelion.festival.exception.WrongBoothId;
-import likelion.festival.exception.WrongCommentId;
-import likelion.festival.exception.WrongPassword;
+import likelion.festival.global.exception.WrongBoothId;
+import likelion.festival.global.exception.WrongCommentId;
+import likelion.festival.global.exception.WrongPassword;
 import likelion.festival.booth.domain.repository.BoothRepository;
 import likelion.festival.comment.domain.repository.CommentRepository;
-import likelion.festival.security.Encrypt;
+import likelion.festival.global.security.Encrypt;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
