+++ src/main/java/likelion/festival/global/config/CorsConfig.java
@@ -1,4 +1,4 @@
-package likelion.festival.config;
+package likelion.festival.global.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
