+++ src/main/java/likelion/festival/global/exception/ExceptionCode.java
@@ -1,4 +1,4 @@
-package likelion.festival.exception;
+package likelion.festival.global.exception;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
