+++ src/main/java/likelion/festival/global/exception/WrongBoothId.java
@@ -1,4 +1,4 @@
-package likelion.festival.exception;
+package likelion.festival.global.exception;
 
 public class WrongBoothId extends IllegalArgumentException{
 }
