+++ src/main/java/likelion/festival/global/exception/WrongMenuId.java
@@ -1,4 +1,4 @@
-package likelion.festival.exception;
+package likelion.festival.global.exception;
 
 public class WrongMenuId extends IllegalArgumentException{
 }
