+++ src/main/java/likelion/festival/global/exception/WrongPassword.java
@@ -1,4 +1,4 @@
-package likelion.festival.exception;
+package likelion.festival.global.exception;
 
 public class WrongPassword extends IllegalArgumentException{
 }
