+++ src/main/java/likelion/festival/global/image/domain/repository/ImageRepository.java
@@ -1,6 +1,6 @@
-package likelion.festival.repository;
+package likelion.festival.global.image.domain.repository;
 
-import likelion.festival.entity.Image;
+import likelion.festival.global.image.domain.Image;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ImageRepository extends JpaRepository<Image, Long> {
