+++ src/main/java/likelion/festival/global/security/MD5Generator.java
@@ -1,4 +1,4 @@
-package likelion.festival.util;
+package likelion.festival.global.security;
 
 import java.io.UnsupportedEncodingException;
 import java.security.MessageDigest;
