+++ src/main/java/likelion/festival/like/application/LikesService.java
@@ -3,8 +3,8 @@ package likelion.festival.like.application;
 import likelion.festival.like.application.dto.LikesResponseDto;
 import likelion.festival.booth.domain.Booth;
 import likelion.festival.like.domain.Likes;
-import likelion.festival.exception.WrongBoothId;
-import likelion.festival.exception.WrongLikesKey;
+import likelion.festival.global.exception.WrongBoothId;
+import likelion.festival.global.exception.WrongLikesKey;
 import likelion.festival.booth.domain.repository.BoothRepository;
 import likelion.festival.like.domain.repository.LikesRepository;
 import lombok.RequiredArgsConstructor;
