+++ src/main/java/likelion/festival/menu/application/MenuService.java
@@ -4,8 +4,8 @@ import likelion.festival.menu.application.dto.MenuRequestDto;
 import likelion.festival.menu.application.dto.MenuResponseDto;
 import likelion.festival.booth.domain.Booth;
 import likelion.festival.menu.domain.Menu;
-import likelion.festival.exception.WrongBoothId;
-import likelion.festival.exception.WrongMenuId;
+import likelion.festival.global.exception.WrongBoothId;
+import likelion.festival.global.exception.WrongMenuId;
 import likelion.festival.booth.domain.repository.BoothRepository;
 import likelion.festival.menu.domain.repository.MenuRepository;
 import lombok.RequiredArgsConstructor;
